Prefabs, variables...
He instaciado la bala, todavia sin funcionalidad, tengo que meterle un script, intentar hacer las interfaces un poco curiosa, y ajustar el comportamiento de las colisiones, partículas y meter sonidos

+++ PR_ZAXXON_delaPenaElisa/Assets/Scripts/Juego/GameManager.cs
@@ -7,7 +7,7 @@ public class GameManager : MonoBehaviour
     //variables y métodos estáticos (públicas)
 
     public static float volume=20f;
-    public static int highscore=0;
+    public static float highscore=0;
     public static float punt = 0;
     public static float speed = 7f;
